sam's feedback
