+++ client/coral-admin/src/containers/Configure/ModerationSettings.js
@@ -12,6 +12,10 @@ const updateModeration = (updateSettings, mod) => () => {
   updateSettings({moderation});
 };
 
+const updateEmailConfirmation = (updateSettings, verify) => () => {
+  updateSettings({requireEmailConfirmation: !verify});
+};
+
 const ModerationSettings = ({settings, updateSettings, onChangeWordlist}) => {
 
   // just putting this here for shorthand below
@@ -20,6 +24,19 @@ const ModerationSettings = ({settings, updateSettings, onChangeWordlist}) => {
 
   return (
     <div>
+      <Card className={`${styles.configSetting} ${settings.requireEmailConfirmation ? on : off}`}>
+        <div className={styles.action}>
+          <Checkbox
+            onChange={updateEmailConfirmation(updateSettings, settings.requireEmailConfirmation)}
+            checked={settings.requireEmailConfirmation} />
+        </div>
+        <div className={styles.content}>
+          <div className={styles.settingsHeader}>{lang.t('configure.require-email-verification')}</div>
+          <p className={settings.requireEmailConfirmation ? '' : styles.disabledSettingText}>
+            {lang.t('configure.require-email-verification-text')}
+          </p>
+        </div>
+      </Card>
       <Card className={`${styles.configSetting} ${settings.moderation === 'PRE' ? on : off}`}>
         <div className={styles.action}>
           <Checkbox
